+++ framework/monolog/monolog/src/Monolog/Formatter/LineFormatter.php
@@ -11,8 +11,6 @@
 
 namespace Monolog\Formatter;
 
-use Monolog\Logger;
-
 /**
  * Formats incoming records into a one-line string
  *
@@ -28,7 +26,7 @@ class LineFormatter extends NormalizerFormatter
     protected $format;
 
     /**
-     * @param string $format The format of the message
+     * @param string $format     The format of the message
      * @param string $dateFormat The format of the timestamp: one supported by DateTime::format
      */
     public function __construct($format = null, $dateFormat = null)
@@ -74,6 +72,17 @@ class LineFormatter extends NormalizerFormatter
             return var_export($data, true);
         }
 
+        if ($data instanceof \Exception) {
+            $previousText = '';
+            if ($previous = $data->getPrevious()) {
+                do {
+                    $previousText .= ', '.get_class($previous).': '.$previous->getMessage().' at '.$previous->getFile().':'.$previous->getLine();
+                } while ($previous = $previous->getPrevious());
+            }
+
+            return '[object] ('.get_class($data).': '.$data->getMessage().' at '.$data->getFile().':'.$data->getLine().$previousText.')';
+        }
+
         return parent::normalize($data);
     }
 
@@ -83,10 +92,11 @@ class LineFormatter extends NormalizerFormatter
             return (string) $data;
         }
 
+        $data = $this->normalize($data);
         if (version_compare(PHP_VERSION, '5.4.0', '>=')) {
-            return json_encode($this->normalize($data), JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE);
+            return $this->toJson($data);
         }
 
-        return stripslashes(json_encode($this->normalize($data)));
+        return str_replace('\\/', '/', json_encode($data));
     }
 }