+++ framework/monolog/monolog/src/Monolog/Formatter/NormalizerFormatter.php
@@ -11,8 +11,6 @@
 
 namespace Monolog\Formatter;
 
-use Monolog\Logger;
-
 /**
  * Normalizes incoming records to remove objects/resources so it's easier to dump to various targets
  *
@@ -72,15 +70,28 @@ class NormalizerFormatter implements FormatterInterface
             return $data->format($this->dateFormat);
         }
 
+        if (is_object($data)) {
+            return sprintf("[object] (%s: %s)", get_class($data), $this->toJson($data, true));
+        }
+
         if (is_resource($data)) {
             return '[resource]';
         }
 
-        return sprintf("[object] (%s: %s)", get_class($data), $this->toJson($data));
+        return '[unknown('.gettype($data).')]';
     }
 
-    protected function toJson($data)
+    protected function toJson($data, $ignoreErrors = false)
     {
+        // suppress json_encode errors since it's twitchy with some inputs
+        if ($ignoreErrors) {
+            if (version_compare(PHP_VERSION, '5.4.0', '>=')) {
+                return @json_encode($data, JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE);
+            }
+
+            return @json_encode($data);
+        }
+
         if (version_compare(PHP_VERSION, '5.4.0', '>=')) {
             return json_encode($data, JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE);
         }